name of classes is changed and some functionality is added.
